requirements.txt
Django==1.6.11
django-appconf==1.0.1
django-model-utils==2.3.1
django-users2==0.1.14
psycopg2
